Implementing a basic drawing loop using Skia.
This adds the UI and logic for drawing the most recent
20 seconds' worth of EEG data from a provided IEEGData
object, and demonstrates it using an EEGDataGenerator (which
currently just produces random data at a configurable interval).

This commit also introduces logging and a (currently empty)
unit test project. An about page has been created as the second
tab in the user interface.