Initial pass at integrating SimplePacketParser.
This creates an implementation of IEEGData that uses the SimplePacketParser. It kicks off a task to continuously read from the serial stream and feed bytes to the parser, raising an event when a complete packet has been received. In the initial pass, it only holds data for a single channel, but will be expanded to hold an arbitrary number.